Move hero content into own component
