+++ src/pages/index.js
@@ -1,20 +1,16 @@
 import React, { useMemo } from 'react';
 import styled from '@emotion/styled';
 import { Helmet } from 'react-helmet';
-import Card from '../components/dev-hub/card';
 import Layout from '../components/dev-hub/layout';
 import Notification from '../components/dev-hub/notification';
-import { H1, SubHeader } from '../components/dev-hub/text';
-import { screenSize, size } from '../components/dev-hub/theme';
-import Button from '../components/dev-hub/button';
 import homepageBackground from '../images/1x/homepage-background.png';
 import useTwitchApi from '../hooks/use-twitch-api';
 import { useSiteMetadata } from '../hooks/use-site-metadata';
-import { getFeaturedCardFields } from '../utils/get-featured-card-fields';
 import {
     AcademiaFeature,
     CommunityFeature,
     EventsFeature,
+    Hero,
     TwitchFeature,
 } from '../components/pages/home';
 
@@ -22,60 +18,8 @@ const BackgroundImage = styled('div')`
     background-image: url(${homepageBackground});
     background-size: cover;
 `;
-const Hero = styled('header')`
-    color: ${({ theme }) => theme.colorMap.devWhite};
-    padding: ${size.xlarge} ${size.large};
-    @media ${screenSize.upToMedium} {
-        padding: ${size.large} ${size.medium};
-    }
-    text-align: center;
-`;
-const Heading = styled(H1)`
-    max-width: 920px;
-    margin: ${size.default} auto;
-    word-wrap: break-word;
-`;
-const Sub = styled(SubHeader)`
-    margin: ${size.default} 0;
-`;
-const CardGallery = styled('section')`
-    display: flex;
-    justify-content: center;
-    margin: ${size.default} ${size.xlarge} ${size.large};
-    @media ${screenSize.upToLarge} {
-        flex-wrap: wrap;
-    }
-    @media ${screenSize.upToMedium} {
-        margin: ${size.default};
-    }
-`;
-const StyledTopCard = styled(Card)`
-    width: 100%;
-    @media ${screenSize.upToLarge} {
-        flex-basis: 50%;
-    }
-    @media ${screenSize.upToMedium} {
-        flex-basis: 100%;
-    }
-`;
 
-// TODO: Generalize as new content types are supported
-const FeaturedHomePageItem = ({ item }) => {
-    if (item.type === 'article') {
-        const { image, slug, title } = getFeaturedCardFields(item);
-        return (
-            <StyledTopCard
-                maxTitleLines={3}
-                image={image}
-                to={slug}
-                title={title}
-                key={title}
-            />
-        );
-    }
-};
-
-export default ({ pageContext: { featuredItems } }) => {
+const Index = ({ pageContext: { featuredItems } }) => {
     const { stream, videos } = useTwitchApi();
     const { title } = useSiteMetadata();
     const twitchVideo = useMemo(() => {
@@ -96,24 +40,7 @@ export default ({ pageContext: { featuredItems } }) => {
                 {stream && (
                     <Notification link={stream.url} title={stream.title} />
                 )}
-                <Hero>
-                    <Heading>
-                        {`ideas.find({"attributes":`}
-                        <br />
-                        {`["fast", "innovative", "original"]})`}
-                    </Heading>
-                    <Sub>What will you create today?</Sub>
-                    <CardGallery>
-                        {featuredItems.map(item => (
-                            <FeaturedHomePageItem item={item} />
-                        ))}
-                    </CardGallery>
-                    <div>
-                        <Button to="/learn" primary>
-                            Learn MongoDB
-                        </Button>
-                    </div>
-                </Hero>
+                <Hero featuredItems={featuredItems} />
                 <TwitchFeature twitchVideo={twitchVideo} />
                 <EventsFeature />
                 <AcademiaFeature />
@@ -122,3 +49,5 @@ export default ({ pageContext: { featuredItems } }) => {
         </Layout>
     );
 };
+
+export default Index;
